requirements.txt
numpy==1.19.1
pyqtgraph==0.11.0
pyserial==3.4
PyQt5==5.15.0